requirements.txt
flask
flask-cors
supabase
python-dotenv
paynow
pydantic

# testing
pytest
